+++ StripePayments/StripePayments/Source/API Bindings/Models/PaymentMethods/STPPaymentMethodEnums.swift
@@ -76,6 +76,8 @@ import Foundation
     case amazonPay
     /// An Alma payment method
     case alma
+    /// A Sunbit payment method
+    case sunbit
     /// A MobilePay payment method
     case mobilePay
     /// A Konbini payment method
@@ -165,6 +167,8 @@ import Foundation
             return "Amazon Pay"
         case .alma:
             return "Alma"
+        case .sunbit:
+            return "Sunbit"
         case .mobilePay:
             return "MobilePay"
         case .konbini:
@@ -256,6 +260,8 @@ import Foundation
             return "amazon_pay"
         case .alma:
             return "alma"
+        case .sunbit:
+            return "sunbit"
         case .mobilePay:
             return "mobilepay"
         case .konbini: